requirements.txt
numpy
pillow
requests
matplotlib
